Update for 2019-2020 season
HTML files were reindented; the org membership was updated; and
Phoenix III content was added.

Note that there are still some "old" images being used, and the YouTube
promo was not updated to 2019. That's because the 2018 video still looks
much better than the 2019 one, and we haven't taken a 2019-2020 group
photo yet.